+++ llama_flutter/lib/Services/utils.dart
@@ -1,6 +1,8 @@
 import 'dart:io';
 
 import 'package:flutter/services.dart';
+import 'package:geocoding/geocoding.dart';
+import 'package:location/location.dart' as loc;
 
 const MethodChannel _platform = MethodChannel('jarvis_data');
 
@@ -57,6 +59,25 @@ class EventManager {
           .toList();
     } on PlatformException catch (e) {
       throw Exception("Failed to load upcoming events: ${e.message}");
+    }
+  }
+
+  Future<String> addInfoEventsAndReminders(String message) async {
+    String lowerMessage = message.toLowerCase();
+    final EventManager _eventManager = EventManager();
+
+    final keywords = ['looking'];
+    if (keywords.any((keyword) => lowerMessage.contains(keyword))) {
+      List<Map<String, dynamic>> reminders =
+          await _eventManager.loadReminders();
+
+      List<Map<String, dynamic>> events =
+          await _eventManager.loadUpcomingEvents();
+
+      return lowerMessage +
+          " here is some information from my calendar for the upcoming year - $events and my due reminders - $reminders";
+    } else {
+      return " ";
     }
   }
 }
@@ -159,13 +180,47 @@ class HomeManager {
       callback(false);
     }
   }
+}
 
-  // Send the summary back to the AI model
-  static void _sendSummaryToAI(String summary, [Function(String)? onSummary]) {
-    // This could involve sending it to a server, logging it, or any other mechanism depending on your AI model setup
-    print("Summary sent to AI: $summary");
-    if (onSummary != null) {
-      onSummary(summary);
+class LocationService {
+  String? _locationAddress;
+
+  Future<String?> fetchAndStoreLocation() async {
+    loc.Location location = loc.Location(); // Use the alias here
+    bool _serviceEnabled;
+    loc.PermissionStatus _permissionGranted;
+
+    _serviceEnabled = await location.serviceEnabled();
+    if (!_serviceEnabled) {
+      _serviceEnabled = await location.requestService();
+      if (!_serviceEnabled) {
+        return null;
+      }
+    }
+
+    _permissionGranted = await location.hasPermission();
+    if (_permissionGranted == loc.PermissionStatus.denied) {
+      _permissionGranted = await location.requestPermission();
+      if (_permissionGranted != loc.PermissionStatus.granted) {
+        return null;
+      }
+    }
+
+    loc.LocationData locationData = await location.getLocation();
+
+    try {
+      List<Placemark> placemarks = await placemarkFromCoordinates(
+        locationData.latitude!,
+        locationData.longitude!,
+      );
+
+      Placemark place = placemarks[0];
+      _locationAddress =
+          "${place.locality}, ${place.postalCode}, ${place.country}";
+      return _locationAddress;
+    } catch (e) {
+      print('Failed to get address: $e');
+      return 'Address unavailable';
     }
   }
 }
