+++ llama_flutter/lib/pages/VoiceView.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:speech_to_text/speech_to_text.dart' as stt;
 
-import '../Services/location.dart';
 import '../Services/udp.dart';
 import '../Services/utils.dart';
